Complete Project with requirements. Pending Javadoc additions.

+++ src/main/java/task1/software1_c482_qkm2_task1/MainController.java
@@ -30,8 +30,8 @@ public class MainController implements Initializable {
     public Button onModifyPart;
     public Button onDeletePart;
     public TableView productsList;
-    public TableColumn col_partID_product;
-    public TableColumn col_partName_product;
+    public TableColumn col_productID;
+    public TableColumn col_productName;
     public TableColumn col_inventoryLevel_product;
     public TableColumn col_priceCostPerUnit_product;
     public TextField filterProduct;
@@ -97,11 +97,11 @@ public class MainController implements Initializable {
         stage.show();
     }
 
-    public Product selectedProduct;
+    public static Product selectedProduct;
     public void setSelectedProduct(Product selected){
         selectedProduct = selected;
     }
-    public Product getSelectedProduct(){
+    public static Product getSelectedProduct(){
         return selectedProduct;
     }
 
@@ -131,6 +131,10 @@ public class MainController implements Initializable {
         String filter = filterPart.getText();
         ObservableList<Part> filteredList = FXCollections.observableArrayList();
         boolean isnull = false;
+        boolean idFound = false;
+        partsList.getSelectionModel().clearSelection();
+        ObservableList<Part> passover = FXCollections.observableArrayList();
+        passover = Inventory.getAllParts();
 
 
         try {
@@ -138,6 +142,8 @@ public class MainController implements Initializable {
             filteredList.add(IdSearch);
             if (IdSearch == null){
                 isnull = true;
+            }else {
+                idFound = true;
             }
         } catch (Exception parseString){
                 filteredList = Inventory.lookupPart(filter);
@@ -148,8 +154,18 @@ public class MainController implements Initializable {
             alert.setContentText("No parts have been found. please refine search criteria.");
             alert.showAndWait();
             filterPart.setText("");
+        }else if (idFound){
+            partsList.setItems(Inventory.getAllParts());
+            for (int i = 0; i < passover.size(); i++){
+                if (String.valueOf(passover.get(i).getId()).contains(filter)){
+                    partsList.getSelectionModel().select(i);
+                    partsList.scrollTo(i);
+                }
+            }
+
         }else {
             partsList.setItems(filteredList);
+
         }
 
         if (filter.isEmpty() || filter.isBlank()){
@@ -162,12 +178,18 @@ public class MainController implements Initializable {
         String filter = filterProduct.getText();
         ObservableList<Product> filteredList = FXCollections.observableArrayList();
         boolean isnull = false;
+        boolean idFound = false;
+        productsList.getSelectionModel().clearSelection();
+        ObservableList<Product> passover = FXCollections.observableArrayList();
+        passover = Inventory.getAllProduct();
 
         try {
             Product IdSearch = Inventory.lookupProduct(Integer.parseInt(filter));
             filteredList.add(IdSearch);
             if (IdSearch == null) {
                 isnull = true;
+            } else{
+                idFound = true;
             }
         } catch (Exception parseString){
             filteredList = Inventory.lookupProduct(filter);
@@ -179,6 +201,15 @@ public class MainController implements Initializable {
             alert.setContentText("No Products have been found. please refine search criteria.");
             alert.showAndWait();
             filterPart.setText("");
+        }else if (idFound) {
+            productsList.setItems(Inventory.getAllProduct());
+            for (int i = 0; i < passover.size(); i++){
+                if (String.valueOf(passover.get(i).getId()).contains(filter)){
+                    productsList.getSelectionModel().select(i);
+                    productsList.scrollTo(i);
+                }
+            }
+
         }else {
             productsList.setItems(filteredList);
         }
@@ -192,17 +223,25 @@ public class MainController implements Initializable {
 
         boolean response;
 
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle("");
-        alert.setHeaderText("Are you sure you want to delete the selected part?");
-        alert.setContentText("Choose Ok to delete or Cancel to keep the part.");
-        ButtonType result = alert.showAndWait().orElse(ButtonType.CANCEL);
-        response = result == ButtonType.OK;
+        Part selected = (Part) partsList.getSelectionModel().getSelectedItem();
+        if (selected == null) {
+            Alert error = new Alert(Alert.AlertType.ERROR);
+            error.setContentText("No part has been selected. Please select a part to delete.");
+            error.showAndWait();
+            filterPart.setText("");
+        }else {
 
-        if (response){
-            Part selected = (Part) partsList.getSelectionModel().getSelectedItem();
-            Inventory.deletePart(selected);
-            partsList.setItems(Inventory.getAllParts());
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("");
+            alert.setHeaderText("Are you sure you want to delete the selected part?");
+            alert.setContentText("Choose Ok to delete or Cancel to keep the part.");
+            ButtonType result = alert.showAndWait().orElse(ButtonType.CANCEL);
+            response = result == ButtonType.OK;
+
+            if (response) {
+                Inventory.deletePart(selected);
+                partsList.setItems(Inventory.getAllParts());
+            }
         }
     }
 
@@ -210,48 +249,91 @@ public class MainController implements Initializable {
 
         boolean response;
 
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle("");
-        alert.setHeaderText("Are you sure you want to delete the selected product?");
-        alert.setContentText("Choose Ok to delete or Cancel to keep the product.");
-        ButtonType result = alert.showAndWait().orElse(ButtonType.CANCEL);
-        response = result == ButtonType.OK;
+        Product selected = (Product) productsList.getSelectionModel().getSelectedItem();
 
-        if (response){
-            Product selected = (Product) productsList.getSelectionModel().getSelectedItem();
-            Inventory.deleteProduct(selected);
-            productsList.setItems(Inventory.getAllProduct());
+        if (selected.getAllAssociatedParts().isEmpty()) {
+
+            if (selected == null) {
+                Alert error = new Alert(Alert.AlertType.ERROR);
+                error.setContentText("No product has been selected. Please select a product to delete.");
+                error.showAndWait();
+                filterProduct.setText("");
+            } else {
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("");
+                alert.setHeaderText("Are you sure you want to delete the selected product?");
+                alert.setContentText("Choose Ok to delete or Cancel to keep the product.");
+                ButtonType result = alert.showAndWait().orElse(ButtonType.CANCEL);
+                response = result == ButtonType.OK;
+
+                if (response) {
+                    Inventory.deleteProduct(selected);
+                    productsList.setItems(Inventory.getAllProduct());
+                }
+            }
+        }else{
+            Alert error = new Alert(Alert.AlertType.ERROR);
+            error.setContentText("Current selected product has associated parts. To delete the parts will need to be disassociated with the product.");
+            error.showAndWait();
+            filterProduct.setText("");
         }
     }
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
+        if (Main.ProgramInitialized.getStatus() == false){
 
-        col_partID_product.setCellValueFactory(new PropertyValueFactory<Product, String>("id"));
-        col_partName_product.setCellValueFactory(new PropertyValueFactory<Product, String>("name"));
-        col_inventoryLevel_product.setCellValueFactory(new PropertyValueFactory<Product, String>("stock"));
-        col_priceCostPerUnit_product.setCellValueFactory(new PropertyValueFactory<Product, String>("price"));
-        productsList.setItems(Inventory.getAllProduct());
+            int id;
 
-        //adds temp parts to model delete before submission.
-        Product four = new Product(1, "TestProd1", 3.50, 1, 3, 6);
-        Product five = new Product(2, "TestProd2", 1.25, 2, 4, 6);
+            col_productID.setCellValueFactory(new PropertyValueFactory<Product, String>("id"));
+            col_productName.setCellValueFactory(new PropertyValueFactory<Product, String>("name"));
+            col_inventoryLevel_product.setCellValueFactory(new PropertyValueFactory<Product, String>("stock"));
+            col_priceCostPerUnit_product.setCellValueFactory(new PropertyValueFactory<Product, String>("price"));
+            productsList.setItems(Inventory.getAllProduct());
 
-        Inventory.getAllProduct().addAll(four, five);
-        // cont delete.
+            id = Main.UniqueIDs.setID();
 
-        col_partID_part.setCellValueFactory(new PropertyValueFactory<Part, String>("id"));
-        col_partName_part.setCellValueFactory(new PropertyValueFactory<Part, String>("name"));
-        col_inventoryLevel_part.setCellValueFactory(new PropertyValueFactory<Part, String>("stock"));
-        col_priceCostPerUnit_part.setCellValueFactory(new PropertyValueFactory<Part, String>("price"));
-        partsList.setItems(Inventory.getAllParts());
+            //adds temp parts to model delete before submission.
+            Product four = new Product(id, "naruto", 3.50, 3, 3, 6);
 
-        //populating test items
-        InHouse one = new InHouse(1, "Test", 3, 4, 5, 1, 55);
-        Outsourced two = new Outsourced(2, "NotReal", 4, 5, 6, 7, "TestComp");
-        InHouse three = new InHouse(3, "Really", 5, 6, 7, 8, 65);
+            id = Main.UniqueIDs.setID();
+            Product five = new Product(id, "Sasuke", 1.25, 5, 4, 6);
 
-        Inventory.getAllParts().addAll(one, two, three);
+            Inventory.getAllProduct().addAll(four, five);
+            // cont delete.
+
+            col_partID_part.setCellValueFactory(new PropertyValueFactory<Part, String>("id"));
+            col_partName_part.setCellValueFactory(new PropertyValueFactory<Part, String>("name"));
+            col_inventoryLevel_part.setCellValueFactory(new PropertyValueFactory<Part, String>("stock"));
+            col_priceCostPerUnit_part.setCellValueFactory(new PropertyValueFactory<Part, String>("price"));
+            partsList.setItems(Inventory.getAllParts());
+
+            //populating test items
+            id = Main.UniqueIDs.setID();
+            InHouse one = new InHouse(id, "Boruto", 3.00, 2, 1, 5, 55);
+
+            id = Main.UniqueIDs.setID();
+            Outsourced two = new Outsourced(id, "Himawari", 4.00, 5, 6, 7, "Child");
+
+            id = Main.UniqueIDs.setID();
+            InHouse three = new InHouse(id, "Sarada", 5.99, 7, 7, 9, 65);
+
+            Inventory.getAllParts().addAll(one, two, three);
+
+            Main.ProgramInitialized.opened = true;
+        }else {
+            col_productID.setCellValueFactory(new PropertyValueFactory<Product, String>("id"));
+            col_productName.setCellValueFactory(new PropertyValueFactory<Product, String>("name"));
+            col_inventoryLevel_product.setCellValueFactory(new PropertyValueFactory<Product, String>("stock"));
+            col_priceCostPerUnit_product.setCellValueFactory(new PropertyValueFactory<Product, String>("price"));
+            productsList.setItems(Inventory.getAllProduct());
+
+            col_partID_part.setCellValueFactory(new PropertyValueFactory<Part, String>("id"));
+            col_partName_part.setCellValueFactory(new PropertyValueFactory<Part, String>("name"));
+            col_inventoryLevel_part.setCellValueFactory(new PropertyValueFactory<Part, String>("stock"));
+            col_priceCostPerUnit_part.setCellValueFactory(new PropertyValueFactory<Part, String>("price"));
+            partsList.setItems(Inventory.getAllParts());
+        }
 
 
     }
